Entity Framework Code First

+++ 03-EntityFramework-Code-First/05-Planck-Constant/Calculation.cs
@@ -8,12 +8,12 @@ namespace _05_Planck_Constant
 {
     class Calculation
     {
-        private static decimal planckConstant = 6.62606896e-34m;
-        private static decimal pi = 3.14159m;
+        private static double planckConstant = 6.62606896e-34; // not decimal !
+        private static double pi = 3.14159;
 
-        public static decimal GetReducedPlanckConstant()
+        public static double GetReducedPlanckConstant()
         {
-            return planckConstant / (2m * pi);
+            return planckConstant / (2 * pi);
         }
     }
 }
